modify bugs in task4

+++ JS/Task2/js/start.js
@@ -3,16 +3,15 @@ function getParams() {
   return postFixPattern.exec(location.href);
 }
 
+function back(params) {
+  sessionStorage.state = JSON.stringify(state);
+  sessionStorage.reStart = "false";
+  location.href = "./task2-log.html" + getParams();
+}
+
 function redirect() {
   //如果返回键被点击，初始化persons中的isAlive的值为1
-  for (var i = 0; i < persons.length; i++) {
-    persons[i].isAlive = 1;
-  }
-  sessionStorage.persons = JSON.stringify(persons);
-  sessionStorage.reStart = "true";
-  sessionStorage.removeItem("logArray");
-  sessionStorage.removeItem("state");
-  location.href = "./task2-log.html";
+  back();
 }
 
 
@@ -69,33 +68,33 @@ for (var i = 0; i < day; i++) {
 }
 
 //给杀手杀人和投票加上记录
-for (i = 0; i < day - 1; i++){
+for (i = 0; i < day - 1; i++) {
   var killLog =
-  '  <div class="prompt__normal">' +
-  '    <div style="font-size:2em">' + logArray[2*i].kill + '号被杀手杀死，真实身份是：' + (logArray[2*i].identity == 0?"杀手":"平民" ) + '</div>' +
-  '  </div>';
+    '  <div class="prompt__normal">' +
+    '    <div style="font-size:2em">' + logArray[2 * i].kill + '号被杀手杀死，真实身份是：' + (logArray[2 * i].identity == 0 ? "杀手" : "平民") + '</div>' +
+    '  </div>';
 
   var voteLog =
-  '  <div class="prompt__normal">' +
-  '    <div style="font-size:2em">' + logArray[2*i+1].kill + '号被投票投死，真实身份是：' + (logArray[2*i+1].identity == 0?"杀手":"平民" )+ '</div>' +
-  '  </div>';
+    '  <div class="prompt__normal">' +
+    '    <div style="font-size:2em">' + logArray[2 * i + 1].kill + '号被投票投死，真实身份是：' + (logArray[2 * i + 1].identity == 0 ? "杀手" : "平民") + '</div>' +
+    '  </div>';
 
   $(".prompt__night").eq(i).after(killLog);
   $(".prompt__last").eq(i).after(voteLog);
 }
 //如果当前步骤在当天的第二步，则单独加上日志
-if(state.step == "2"){
+if (state.step >= "2") {
   killLog =
-  '  <div class="prompt__normal">' +
-  '    <div style="font-size:2em">' + logArray[2*(day-1)].kill + '号被杀手杀死，真实身份是：' + (logArray[2*(day-1)].identity == 0?"杀手":"平民" ) + '</div>' +
-  '  </div>';
+    '  <div class="prompt__normal">' +
+    '    <div style="font-size:2em">' + logArray[2 * (day - 1)].kill + '号被杀手杀死，真实身份是：' + (logArray[2 * (day - 1)].identity == 0 ? "杀手" : "平民") + '</div>' +
+    '  </div>';
   $(".prompt__night").last().after(killLog);
 }
 
 
-  //折叠标签
-  $(".date").next().hide();
-  $(".date").last().next().show();
+//折叠标签
+$(".date").next().hide();
+$(".date").last().next().show();
 
 //添加toggle效果
 for (i = 0; i < day; i++) {
@@ -130,15 +129,14 @@ switch (step - 1) {
 // 点击杀手杀人按钮
 $(".prompt__killer").click(function (e) {
   phase = "kill";
-  if(state.day != Math.ceil(parseInt(event.target.id) / 4)){
+  if (state.day != Math.ceil(parseInt(event.target.id) / 4)) {
     alert("请进行下一步操作！");
     return;
   }
   if (state.step < "1") {
     alert("请按游戏流程操作！");
     return;
-  }
-  else if(state.step > "1"){
+  } else if (state.step > "1") {
     alert("请进行下一步操作！");
     return;
   }
@@ -146,81 +144,92 @@ $(".prompt__killer").click(function (e) {
   $("#day" + day).next().find(".prompt__killer").addClass("prompt__disable");
   sessionStorage.state = JSON.stringify(state);
   sessionStorage.phase = phase;
-  location.href = "./task2-kill.html"
+  location.href = "./task2-kill.html" + getParams();
 });
 
 //亡灵发表遗言
 $(".prompt__death").click(function (e) {
-  if(state.day != Math.ceil(parseInt(event.target.id) / 4)){
+  if (state.day != Math.ceil(parseInt(event.target.id) / 4)) {
     alert("请进行下一步操作！");
     return;
   }
   if (state.step < "2") {
     alert("请按游戏流程操作！");
     return;
-  }
-  else if(state.step > "2"){
+  } else if (state.step > "2") {
     alert("请进行下一步操作！");
     return;
   }
   alert("请死者亮明身份并发表遗言！");
   state.step = ++step + "";
   $("#day" + day).next().find(".prompt__death").addClass("prompt__disable");
+  sessionStorage.state = JSON.stringify(state);
 });
 
 
 //玩家依次发言
 $(".prompt__others").click(function (e) {
-  if(state.day != Math.ceil(parseInt(event.target.id) / 4)){
+  if (state.day != Math.ceil(parseInt(event.target.id) / 4)) {
     alert("请进行下一步操作！");
     return;
   }
   if (state.step < "3") {
     alert("请按游戏流程操作！");
     return;
-  }
-  else if(state.step > "3"){
+  } else if (state.step > "3") {
     alert("请进行下一步操作！");
     return;
   }
   alert("请玩家依次发言！");
   state.step = ++step + "";
   $("#day" + day).next().find(".prompt__others").addClass("prompt__disable");
+  sessionStorage.state = JSON.stringify(state);
 });
 
 //全民投票
 $(".prompt__vote").click(function (e) {
-  if(state.day != Math.ceil(parseInt(event.target.id) / 4)){
+  if (state.day != Math.ceil(parseInt(event.target.id) / 4)) {
     alert("请进行下一步操作！");
     return;
   }
   if (state.step < "4") {
     alert("请按游戏流程操作！");
     return;
-  }
-  else if(state.step > "4"){
+  } else if (state.step > "4") {
     alert("请进行下一步操作！");
     return;
   }
   phase = "vote";
   sessionStorage.phase = phase;
   sessionStorage.state = JSON.stringify(state);
-  location.href = "./task2-kill.html";
+  location.href = "./task2-kill.html" + getParams();
 });
 
 
 //点击法官日志按钮事件
-$(".log").click(function (e) {
-  sessionStorage.state = JSON.stringify(state);
-  sessionStorage.reStart = "false";
-  location.href = "./task2-log.html";
-});
+$(".log").click(back);
 
 
 //结束游戏
 $(".end").click(function (e) {
-  sessionStorage.clear();
-  location.href=  "./task2-assign.html";
+
+  var s = confirm("结束本轮游戏吗?");
+  if (s == true) {
+    sessionStorage.clear();
+    location.href = "./task2-home.html";
+  } else {
+    return;
+  }
+
 
 });
 
+//点击关闭按钮
+$(".close").click(function () {
+  var s = confirm("结束本轮游戏吗?");
+  if (s == true) {
+    location.href = "./task2-home.html";
+  } else {
+    return;
+  }
+});
